WIP: refactor charts model into recipe and food m-to-m
